Closes #5
- The DatacenterSimple.submitCloudletToVm computes the transfer time for Cloudlet files and
  submits the Cloudlet to the scheduler, informing this time. The scheduler was then
  recomputing the time. By this way, the Datacenter added its computed transfer time
  and the scheduler too, doubling the transfer time.
  Now the Datacenter computes the transfer time and passes it to the scheduler when
  submitting the Cloudlet. The cloudlet returns the expected Cloudlet finish time,
  including the transfer time received. The Datacenter so just uses this expected
  finish time without adding again the transfer time.
- To fix this issue, some refactorings were performed in the scheduler classes to
  remove duplicated code and move more code to the CloudletSchedulerAbstract class.
  The CloudletExecutionInfo received more 2 attributes: fileTransferTime and lastProcessingTime
  to correctly implement the transfer time delay without increasing the Cloudlet length
  to give the idea of transfer delay.

+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/Log.java
@@ -17,7 +17,7 @@ import java.lang.management.ManagementFactory;
  *
  * @author Anton Beloglazov
  * @since CloudSim Toolkit 2.0
- * @todo @author manoelcampos Could be replaced by 
+ * @todo @author manoelcampos Could be replaced by
  * {@link http://logging.apache.org/log4j/2.x/ Apache Log4j}
  */
 public class Log {
@@ -26,15 +26,15 @@ public class Log {
      * of a log message.
      */
     public enum Level {
-        INFO, 
+        INFO,
         ERROR,
-        
+
         /**
          * A log level for messages that will be shown only when the application
          * is run in debug mode.
          */
         DEBUG};
-    
+
     /**
      * The Constant LINE_SEPARATOR.
      */
@@ -55,19 +55,19 @@ public class Log {
      * Buffer to avoid creating new string builder upon every print.
      */
     private static final StringBuilder buffer = new StringBuilder();
-    
+
     /**
      * Checks if application is running in debug mode.
      * "jdwp" is the acronym for "Java Debug Wire Protocol" that
      * may exists as an application parameter to define
      * the application is running in debug mode.
-     * 
-     * @see #isDebug() 
+     *
+     * @see #isDebug()
      */
-    private static boolean debug = 
+    private static boolean debug =
             ManagementFactory.getRuntimeMXBean().getInputArguments()
                     .toString().indexOf("jdwp") > 0;
-    
+
     /**
      * Prints a message.
      *
@@ -115,7 +115,7 @@ public class Log {
      */
     public static void printConcat(Object... messages) {
         if (isEnabled()) {
-            buffer.setLength(0); // Clear the buffer		    
+            buffer.setLength(0); // Clear the buffer
             for (int i = 0; i < messages.length; i++) {
                 buffer.append(String.valueOf(messages[i]));
             }
@@ -131,7 +131,7 @@ public class Log {
      */
     public static void printConcatLine(Object... messages) {
         if (isEnabled()) {
-            buffer.setLength(0); // Clear the buffer		    
+            buffer.setLength(0); // Clear the buffer
             for (int i = 0; i < messages.length; i++) {
                 buffer.append(String.valueOf(messages[i]));
             }
@@ -171,7 +171,7 @@ public class Log {
 
     /**
      * Prints a string formated as in String.printFormatted(), followed by a new
-     * line, that will be printed only according to 
+     * line, that will be printed only according to
      * the specified level
      *
      * @param level the level that define the kind of message
@@ -184,7 +184,7 @@ public class Log {
     public static void println(Level level, Class _class, double time, String format, Object... args) {
         if((level == Level.DEBUG && isDebug()) || (level != Level.DEBUG)){
             String msg = String.format(format, args);
-            printFormattedLine("Time %d %s/%s\n   %s", (long)time, level.name(), _class.getSimpleName(), msg);
+            printFormattedLine("Time %.1f %s/%s\n   %s", time, level.name(), _class.getSimpleName(), msg);
         }
     }
 
@@ -253,6 +253,6 @@ public class Log {
     public static boolean isDebug() {
         return debug;
     }
-    
+
 
 }
